chore(gh-pages): get rid of _data dir, site data moved to yml
