+++ pages/11ty/site.config.js
@@ -1,0 +1,11 @@
+const fs = require('fs');
+const path =require('path')
+const yaml = require('js-yaml');
+
+const content = fs.readFileSync(path.resolve(__dirname, '../site.yml'), 'utf8');
+const siteConfig = yaml.load(content, {});
+
+module.exports = (config) => {
+  config.addGlobalData('site', siteConfig);
+};
+Object.assign(module.exports, siteConfig);
