--- pages/views/_data/site.json
@@ -1,19 +1,0 @@
-{
-  "name": "Exadel Smart Library",
-  "description": "ESL - UI web components open-source library",
-  "url": "https://esl-ui.com",
-  "urlPrefix": "",
-  "github": {
-    "url": "https://github.com/exadel-inc/esl",
-    "srcUrl": "https://github.com/exadel-inc/esl/tree/main/"
-  },
-  "rewriteRules": {
-    "docs/INSTALLATION_GUIDE.md": "/introduction",
-    "docs/BROWSER_SUPPORT.md": "/introduction/browser-support",
-    "src/modules/esl-base-element/README.md": "/utils/esl-base-element",
-    "src/modules/esl-media-query/README.md": "/utils/esl-media-query",
-    "src/modules/esl-traversing-query/README.md": "/utils/esl-traversing-query",
-    "src/modules/esl-trigger/README.md": "/components/esl-trigger",
-    "src/modules/esl-toggleable/README.md": "/components/esl-toggleable"
-  }
-}
