+++ pages/views/_includes/navigation/header.njk
@@ -3,7 +3,7 @@
     <a class="header-link header-logo fill-light logo-animated" href="{{ '/' | url }}" aria-label="Landing page">
       {% include 'static/assets/logo.svg' %}
       <span class="header-short-name">esl</span>
-      <span class="header-long-name">{{ site.name }}</span>
+      <span class="header-long-name">{{ site.brand }}</span>
     </a>
     <ul class="header-nav">
       <li class="header-nav-item"><a class="header-link" href="{{ '/examples' | url }}">Examples</a></li>
