+++ pages/views/_includes/navigation/sidebar.njk
@@ -4,7 +4,7 @@
   <nav class="sidebar-nav">
     <a class="sidebar-heading" href="{{ '/' | url }}">
       <span class="sidebar-logo fill-light logo-animated">{% include 'static/assets/logo.svg' %}</span>
-      <span class="sidebar-title">{{ site.name }}</span>
+      <span class="sidebar-title">{{ site.brand }}</span>
     </a>
 
     <div class="sidebar-content">
